Moved AST base to separate crate This is so I can move the parser to its own crate, improving parser compile time issues
