+++ Cargo.toml
@@ -16,3 +16,6 @@ log = "0.3.1"
 
 [dependencies.seax_util]
 git = "https://github.com/hawkw/seax_util.git"
+
+[dependencies.ast_base]
+path = "ast_base"
