+++ ast_base/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "ast_base"
+version = "0.1.0"
+authors = ["Hawk Weisman <hi@hawkweisman.me>"]
