+++ src/ast/mod.rs
@@ -1,3 +1,6 @@
+extern crate ast_base;
+pub use self::ast_base::*;
+
 use seax::cell::SVMCell;
 use seax::cell::Atom::*;
 use seax::cell::Inst::*;
@@ -21,42 +24,28 @@ use std::hash::Hash;
 #[cfg(test)]
 mod tests;
 
-/// Expression
-///
-/// All Seax Scheme expressions are one of the following
-///
-///  + Nested S-Expressions
-///  + Identifiers
-///  + Lists
-///  + Numbers
-///     - signed int
-///     - unsigned int
-///     - floating-point
-///  + Characters
-///  + Strings
-///
-///  TODO: implement the entire Scheme 'numeric tower'
-///  TODO: macros should happen
-///  TODO: figure out quasiquote somehow.
-#[derive(Clone, PartialEq)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub enum ExprNode {
+impl ASTNode for RootNode {
+    #[unstable(feature="compile")]
+    #[allow(unused_variables)]
+    fn compile<'a>(&'a self, state: &'a SymTable<'a>) -> CompileResult {
+        Err("UNINPLEMENTED".to_string())
+    }
+
     #[stable(feature = "ast", since = "0.0.2")]
-    Root(RootNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    SExpr(SExprNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    Name(NameNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    ListConst(ListNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    NumConst(NumNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    BoolConst(BoolNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    StringConst(StringNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    CharConst(CharNode),
+    fn print_level(&self, level: usize) -> String {
+        let mut result = String::new();
+        for expr in &self.exprs {
+            write!(result, "{}", &expr.print_level(level + 1))
+                .unwrap();
+        }
+        result
+    }
+}
+
+impl fmt::Debug for RootNode {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.prettyprint())
+    }
 }
 
 impl ASTNode for ExprNode {
@@ -95,65 +84,6 @@ impl fmt::Debug for ExprNode {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.prettyprint())
     }
-}
-
-#[derive(Clone, PartialEq)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub enum NumNode {
-    #[stable(feature = "ast", since = "0.0.2")]
-    IntConst(IntNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    UIntConst(UIntNode),
-    #[stable(feature = "ast", since = "0.0.2")]
-    FloatConst(FloatNode)
-}
-
-impl fmt::Debug for NumNode {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.prettyprint())
-    }
-}
-
-/// AST node for the root of a program's AST
-#[derive(Clone, PartialEq)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct RootNode { pub exprs: Vec<ExprNode> }
-
-impl fmt::Debug for RootNode {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.prettyprint())
-    }
-}
-
-impl ASTNode for RootNode {
-    #[unstable(feature="compile")]
-    #[allow(unused_variables)]
-    fn compile<'a>(&'a self, state: &'a SymTable<'a>) -> CompileResult {
-        Err("UNINPLEMENTED".to_string())
-    }
-
-    #[stable(feature = "ast", since = "0.0.2")]
-    fn print_level(&self, level: usize) -> String {
-        let mut result = String::new();
-        for expr in &self.exprs {
-            write!(result, "{}", &expr.print_level(level + 1))
-                .unwrap();
-        }
-        result
-    }
-}
-
-/// AST node for an S-expression.
-///
-/// This includes function application, assignment,
-/// function definition, et cetera...Scheme is not a complexl anguage.
-#[derive(Clone, PartialEq)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct SExprNode {
-    #[stable(feature = "ast", since = "0.0.2")]
-    pub operator: Box<ExprNode>,
-    #[stable(feature = "ast", since = "0.0.2")]
-    pub operands: Vec<ExprNode>,
 }
 
 impl ASTNode for SExprNode {
@@ -409,18 +339,6 @@ impl fmt::Debug for SExprNode {
     }
 }
 
-/// AST node for a list literal
-#[derive(Clone, PartialEq)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct ListNode { pub elements: Vec<ExprNode> }
-#[stable(feature = "ast", since = "0.0.4")]
-impl fmt::Debug for ListNode {
-    #[stable(feature = "ast", since = "0.0.4")]
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.prettyprint())
-    }
-}
-
 impl ASTNode for ListNode {
     #[unstable(feature="compile")]
     #[allow(unused_variables)]
@@ -446,53 +364,8 @@ impl ASTNode for ListNode {
 
 }
 
-/// AST node for an identifier
-#[derive(Clone, PartialEq)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct NameNode { pub name: String }
-
-impl NameNode {
-    /// Returns true if this is a keyword
-    #[inline]
-    #[stable(feature = "ast", since = "0.0.3")]
-    fn is_kw(&self) -> bool {
-        match self.name.as_ref() {
-            "access" | "define-syntax" | "macro"  | "and"  | "delay"
-            | "make-environment" | "begin"  | "do"| "named-lambda"
-            | "bkpt" | "fluid-let" | "or" | "case" | "if" | "quasiquote"
-            | "cond" | "in-package" | "quote" | "cons-stream" | "lambda"
-            | "scode-quote" | "declare" | "let" | "sequence" | "default-object?"
-            | "let*" | "set!" | "define" | "let-syntax" | "the-environment"
-            | "define-integrable" | "letrec" | "unassigned?" | "define-macro"
-            | "local-declare" | "using-syntax" | "define-structure" | "car"
-            | "cdr" | "cons" | "nil" | "nil?" | "atom?" => true,
-            _ => false
-        }
-    }
-    /// Returns true if this is an arithmetic operator
-    #[inline]
-    #[stable(feature = "ast", since = "0.0.3")]
-    fn is_arith(&self) -> bool {
-      match self.name.as_ref() {
-         "+" | "-" | "*" | "/" | "%" => true,
-         _ => false
-      }
-   }
-    /// Returns true if this is a comparison operator
-    #[inline]
-    #[stable(feature = "ast", since = "0.0.3")]
-    fn is_cmp(&self) -> bool {
-      match self.name.as_ref() {
-         "=" | "!=" | ">" | "<" | ">=" | "<=" => true,
-         _ => false
-      }
-   }
-
-   #[stable(feature = "ast", since = "0.0.4")]
-   pub fn new(name: String) -> Self { NameNode {name: name} }
-}
 #[stable(feature = "ast", since = "0.0.4")]
-impl fmt::Debug for NameNode {
+impl fmt::Debug for ListNode {
     #[stable(feature = "ast", since = "0.0.4")]
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.prettyprint())
@@ -541,13 +414,14 @@ impl ASTNode for NameNode {
 
 }
 
-/// AST node for an integer constant
-#[derive(Clone, PartialEq,Debug)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct IntNode {
-    #[stable(feature = "ast", since = "0.0.2")]
-    pub value: i64
+#[stable(feature = "ast", since = "0.0.4")]
+impl fmt::Debug for NameNode {
+    #[stable(feature = "ast", since = "0.0.4")]
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.prettyprint())
+    }
 }
+
 
 impl ASTNode for NumNode {
     #[stable(feature="compile",since="0.0.3")]
@@ -585,56 +459,10 @@ impl ASTNode for NumNode {
     }
 }
 
-/// AST node for an unsigned integer constant
-#[derive(Clone, PartialEq,Debug)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct UIntNode {
-    #[stable(feature = "ast", since = "0.0.2")]
-    pub value: u64
-}
-
-/// AST node for a floating-point constant
-#[derive(Clone, PartialEq,Debug)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct FloatNode {
-    #[stable(feature = "ast", since = "0.0.2")]
-    pub value: f64
-}
-
-/// AST node for a boolean constant
-#[derive(Clone, PartialEq,Debug)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct BoolNode {
-    #[stable(feature = "ast", since = "0.0.2")]
-    pub value: bool
-}
-
-impl ASTNode for BoolNode {
-    #[stable(feature="compile", since="0.0.6")]
-    #[allow(unused_variables)]
-    fn compile<'a>(&'a self,state:  &'a SymTable)    -> CompileResult {
-        match self.value {
-            true    => Ok(vec![InstCell(LDC), AtomCell(SInt(1))]),
-            false   => Ok(vec![InstCell(NIL)])
-        }
+impl fmt::Debug for NumNode {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.prettyprint())
     }
-
-    #[stable(feature = "ast", since = "0.0.2")]
-    fn print_level(&self, level: usize) -> String {
-        let mut tab = String::new();
-        for _ in 0 .. level {tab.push_str(INDENT);};
-
-        format!("{}Boolean: {}\n", tab, self.value)
-    }
-}
-
-
-/// AST node for a character constant
-#[derive(Clone, PartialEq,Debug)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct CharNode {
-    #[stable(feature = "ast", since = "0.0.2")]
-    pub value: char
 }
 
 impl ASTNode for CharNode {
@@ -651,12 +479,6 @@ impl ASTNode for CharNode {
         format!("{}Character: \'{}\'\n", tab, self.value)
     }
 }
-
-
-/// AST node for a  string constant
-#[derive(Clone, PartialEq,Debug)]
-#[stable(feature = "ast", since = "0.0.2")]
-pub struct StringNode { pub value: String }
 
 impl ASTNode for StringNode {
     /// Method to compile a String.
@@ -676,5 +498,25 @@ impl ASTNode for StringNode {
     #[allow(unused_variables)]
     fn print_level(&self, level: usize) -> String {
         format!("String: \"{}\"\n", self.value)
+    }
+}
+
+
+impl ASTNode for BoolNode {
+    #[stable(feature="compile", since="0.0.6")]
+    #[allow(unused_variables)]
+    fn compile<'a>(&'a self,state:  &'a SymTable)    -> CompileResult {
+        match self.value {
+            true    => Ok(vec![InstCell(LDC), AtomCell(SInt(1))]),
+            false   => Ok(vec![InstCell(NIL)])
+        }
+    }
+
+    #[stable(feature = "ast", since = "0.0.2")]
+    fn print_level(&self, level: usize) -> String {
+        let mut tab = String::new();
+        for _ in 0 .. level {tab.push_str(INDENT);};
+
+        format!("{}Boolean: {}\n", tab, self.value)
     }
 }
